# specs/001-cloud-event-driven-deployment/contracts/api.yaml
openapi: 3.0.3
info:
  title: Hackathon Todo Evolution API
  description: Backend API for Todo application with event-driven architecture
  version: 0.2.0
  contact:
    name: Todo Evolution Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://todo-api.example.com
    description: Production (Oracle OKE)
...
paths:
...
  /api/tasks:
    get:
      tags: [tasks]
      summary: List tasks
      description: Get paginated list of user's tasks with filtering
      operationId: listTasks
      security:
        - BearerAuth: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: search
          in: query
          description: Search in title
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, all]
            default: all
        - name: tags
          in: query
          description: Comma-separated tags
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [due_date, priority, created_at, title]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [tasks]
      summary: Create task
      description: Create a new task with optional recurrence
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Get task
      description: Get a specific task by ID
      operationId: getTask
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [tasks]
      summary: Update task
      description: Partially update a task
      operationId: updateTask
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Concurrent modification conflict

    delete:
      tags: [tasks]
      summary: Delete task
      description: Delete a task
      operationId: deleteTask
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # Notification Endpoints
  # =============================================================================
...
  /events/tasks:
    post:
      tags: [events]
      summary: Handle task events
      description: Dapr delivers task events from Kafka to this endpoint
      operationId: handleTaskEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
      responses:
        '200':
          description: Event processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: SUCCESS
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Better Auth JWT token
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Complete Phase V
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: Implement cloud deployment with Dapr and Kafka
        completed:
          type: boolean
          default: false
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        due_date:
          type: string
          format: date-time
          nullable: true
          example: '2025-02-15T23:59:59Z'
        reminder_at:
          type: string
          format: date-time
          nullable: true
          example: '2025-02-15T22:00:00Z'
        tags:
          type: array
          items:
            type: string
          maxItems: 10
          example: ["hackathon", "urgent"]
        recurrence_rule:
          type: string
          nullable: true
          description: iCal RRULE format
          example: FREQ=WEEKLY;BYDAY=FR
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        due_date:
          type: string
          format: date-time
          nullable: true
        reminder_at:
          type: string
          format: date-time
          nullable: true
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        recurrence_rule:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
          description: For optimistic locking
    TaskRead:
      type: object
      properties:
        id:
          type: integer
          example: 123
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: Complete Phase V
        description:
          type: string
          nullable: true
        completed:
          type: boolean
          example: false
        priority:
          type: string
          enum: [low, medium, high]
          example: high
        due_date:
          type: string
          format: date-time
          nullable: true
        reminder_at:
          type: string
          format: date-time
          nullable: true
        tags:
          type: array
          items:
            type: string
          example: ["hackathon", "urgent"]
        recurrence_rule:
          type: string
          nullable: true
        parent_task_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskRead'
        total:
          type: integer
          example: 42
...
    Error:
      type: object
      properties:
        detail:
          type: string
          example: Invalid request data
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'